refactor: move provider, and common to pkg package. minor fixes
