+++ pkg/provider/vault/client_logger.go
