+++ pkg/provider/vault/vault.go
@@ -25,8 +25,8 @@ import (
 	"github.com/bank-vaults/internal/injector"
 	"github.com/bank-vaults/vault-sdk/vault"
 
-	"github.com/bank-vaults/secret-init/common"
-	"github.com/bank-vaults/secret-init/provider"
+	"github.com/bank-vaults/secret-init/pkg/common"
+	"github.com/bank-vaults/secret-init/pkg/provider"
 )
 
 const ProviderName = "vault"
