+++ pkg/provider/vault/vault_test.go
